+++ tools/generator/test/PBXProj+Testing.swift
@@ -1,0 +1,9 @@
+@testable import XcodeProj
+
+extension PBXProj {
+    /// Converts temporary references into stable ones, for equality asserts.
+    func fixReferences() throws {
+        let referenceGenerator = ReferenceGenerator(outputSettings: .init())
+        try referenceGenerator.generateReferences(proj: self)
+    }
+}
